Model::find(); if (!Model) { redirect } for all Controller

+++ app/Http/Controllers/ParticipantsController.php
@@ -46,7 +46,9 @@ class ParticipantsController extends Controller
 
     public function edit($id)
     {
-        $participant = Participant::with('uuidCard', 'club')->findOrFail($id);
+        $participant = Participant::with('uuidCard', 'club')->find($id);
+
+        if (!$participant) {  return redirect()->route('participants.index')->with('error', 'You cannot edit the participant with id ' . $id . ' becouse not exist! '); }
 
         $clubs = Club::all();
         $uuidList = UuidCard::all();
@@ -62,7 +64,9 @@ class ParticipantsController extends Controller
             'name' => 'required',
         ]);
 
-        $participant = Participant::findOrFail($id);
+        $participant = Participant::find($id);
+
+        if (!$participant) {  return redirect()->route('participants.index')->with('error', 'You cannot update the participant with id ' . $id . ' becouse not exist! '); }
 
         $participant->update([
             'club_id' => $request->input('club_id'),
@@ -77,7 +81,9 @@ class ParticipantsController extends Controller
     {
         $participant = Participant::with(['participantManagers' => function ($query) {
             $query->with('stage', 'uuidCard');
-        }])->findOrFail($id);
+        }])->find($id);
+
+        if (!$participant) {  return redirect()->route('participants.index')->with('error', 'You cannot manage the participant with id ' . $id . ' becouse not exist! '); }
 
         $categories = Category::all();
 
@@ -108,7 +114,9 @@ class ParticipantsController extends Controller
             'manage.*.post_finish' => 'required|date_format:H:i:s',
         ]);
 
-        $participant = Participant::findOrFail($id);
+        $participant = Participant::find($id);
+
+        if (!$participant) {  return redirect()->route('participants.index')->with('error', 'You cannot update the participant manage with id ' . $id . ' becouse not exist! '); }
 
         foreach ($request->input('manage') as $manageId => $manage) {
             $manager = $participant->participantManagers()->findOrFail($manageId);
@@ -137,7 +145,10 @@ class ParticipantsController extends Controller
 
     public function destroy($id)
     {
-        $participant = Participant::findOrFail($id);
+        $participant = Participant::find($id);
+
+        if (!$participant) {  return redirect()->route('participants.index')->with('error', 'You cannot remove the participant with id ' . $id . ' becouse not exist! '); }
+
         $participant->delete();
 
         return redirect()->route('participants.index')->with('success', $participant->name . ' has been deleted!');
